3.1.7 Zmiana nazwy pakietu
