+++ src/main/java/edu/kis/vh/nursery/stack/IntArrayStack.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.stack;
 
 import edu.kis.vh.nursery.Stack;
